requirements.txt
python-dotenv
requests
sentence-transformers
faiss-cpu
uvicorn
fastapi
streamlit
